Refactor: Move isAbsoluteUrl function to a separate file in the utils folder for better reusability.
- Move addLocaleToUrl, isPlanetDomain to  the utils folder.
